tabulate tail-end pending, improve formatting, fix bugs
If, at the end of a trace, a sys_enter was seen for a task, but a sys_exit
was not, the syscall is considered "pending".  The elapsed time in that
syscall is tabulated and added to the total elapsed time for that syscall,
although the count, min, max, and average are not influenced.

Similarly, if at the end of a trace, a task is found to still be in user
mode, the time from the last event to the end of the trace is tabulated
and added to user time for that task.

Idle time is similar, although we do need to know if the idle time should
be also accounted for with respect to the elapsed time of a pending
syscall, if a syscall was active at the time the task went idle.

Signed-off-by:  Paul A. Clarke <pc@us.ibm.com>