discover PIDs later in trace
Some tasks will first appear in events like "sched_switch", which will
likely be an event _for_ a different task, often task 0 (idle task).
Because of this, there is no way to determine the task's PID until
another event _for_ that task is processed.

Set the PID to 'unknown' in these cases, and change any 'unknown'
PIDs to actual PIDs when subsequent events for that task are processed.

Fixes #1
Fixes #19

Also, fix a bug where the wrong variable name was used.

Signed-off-by: Paul A. Clarke <pc@us.ibm.com>