correct output alignment
Some output fields were improperly aligned, because an unnecessary extra
space was added for continued print statements, like:

	print "a",
	print "b"

Here, a space is not needed before "b" in order to separate "a" and "b",
as Python adds one automatically.

Also, some lines are now indented with tabs instead of trying to accommodate
the timestamp (with a repeated timestamp or whitespace), just for
simplicity.

Finally, some trailing whitespace was removed in the code.

Signed-off-by: Paul A. Clarke <pc@us.ibm.com>