handle out-of-order events
The "perf" command records events in the perf.data file as they come.
Unfortunately, they may come with timestamps out-of-order.  This will
confuse a tool like curt which attempts to track state changes as the
events are processed, in file order, presuming that file order is also
temporal order.  A sequence like the following cannot be handled in
file order:

	syscall_enter
	hcall_enter
	syscall_exit
	hcall_exit

The solution provided here is to enqueue events in a queue which is
maintained in order, sorted by timestamp.  Then, the events are
processed from the head of the queue, in time order.

The depth of the queue is configurable with a new "--window N"
command-line option.  The default depth is 20 events.

Note: this has allowed for some simplification and better design of
the code at the expense of some performance impact.

Signed-off-by: Paul A. Clarke <pc@us.ibm.com>