performance improvements, cross-platform fixes
Significant redesign.

Added aggregates in the form of classes to eliminate the use of multi-level
dictionaries.  Fixes #5

Added support for run-time determination of syscall numbers to correct use
of synthetic events for clone and exec syscalls.  Fixes #2

Also refactored a lot of code into class methods to reduce the code size and
put common code into a single place for ease of maintenance.

Signed-off-by:  Paul A. Clarke <pc@us.ibm.com>