support older perf API by default
Most current distributions support a version of perf with an older,
and slightly different API for the event handling routines.  The older
API does not have a final "perf_sample_dict" parameter.  This is used
by curt.py to determine the tasks' process IDs.

If the current version of perf is used with the current version of
curt, then an error occurs:
  TypeError: powerpc__hcall_entry_new() takes exactly 10 arguments (9 given)

A workaround is to just abandon the use of process IDs, and classify all
tasks under an 'unknown' PID.

A new command line parameter, "--api=<n>" is added by this patch.

To use curt with the older perf API:

$ perf script -s ./curt.py -- --api=1

To use curt with the newer perf API:

$ perf script -s ./curt.py -- --api=2

The default behavior is now changed to use the older perf API, so that
curt.py can be used easily on current Linux distributions.

Fixes #7

Signed-off-by:  Paul A. Clarke <pc@us.ibm.com>