add "record" capability
The perf record command required to generate a perf data file with the
necessary events is long and complex.  Plus, it's an extra step.

Add the ability to record the required data using a simple invocation
of curt, then proceed to generate the report all in one step.

$ ./curt.py --record --all ./command

Additionally, provide subsets of events which can be combined to
generate custom recordings with less than *all* events.

$ ./curt.py --record syscalls,hcalls,sched ./command

Finally, update and improve the help text.

Signed-off-by:  Paul A. Clarke <pc@us.ibm.com>