use parent state to initialize new child task
Currently, when a new task is created, curt will synthesize
a CLONE system call entry to set the initial state of the task.
Unfortunately, this requires knowing the numeric code of the
`clone` system call, which is architecture-dependent and for
which there is no easy mapping from the name.  An exhaustive
search is done for every likely valid numeric code (and then
some).

A better approach is to "clone" the state from the parent
task, because it is indeed in a `clone` system call itself,
so no search is required for the code, as it can be simply
extracted from the parent.

The new edge case that must be handled is when the first
event for the _parent_ is a `sched_process_fork`.  In this case,
a `Task` must be created and initialized for the parent task.

Signed-off-by: Paul A. Clarke <pc@us.ibm.com>