use previous state to initialize newly exec'd task state
Currently, when a task's image is replaced via `exec()`, curt
will synthesize an `execve` system call entry to set the initial
state of the task.

Unfortunately, this requires knowing the numeric code of the
`execve` system call, which is architecture-dependent and for
which there is no easy mapping from the name.  An exhaustive
search is done for every likely valid numeric code (and then
some).

A better approach is to copy the previous state from the
task, because it is indeed in a `execve` system call,
so no search is required for the numeric code, as it can be
simply extracted from the previous state.

Signed-off-by: Paul A. Clarke <pc@us.ibm.com>